move validate_buffer_config into new replay buffer utils file
